refactor : BigDecimal로 리팩토링

+++ src/main/java/com/umc/finly/domain/market/dto/MarketIndexResponse.java
@@ -5,14 +5,16 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class MarketIndexResponse {
 
-    private Double kospi;
-    private Double kosdaq;
+    private BigDecimal kospi;
+    private BigDecimal kosdaq;
     private Integer fearGreed;
     private String fearGreedStatus;
     private String updatedAt;
